requirements.txt
dblpy
websockets
aiohttp
PyNaCl
multidict
async-timeout
attrs
pycparser
pymongo
motor
dnspython
aiohttp
asyncio
psutil
timeago
numpy
Augmentor
jishaku
requests
colorthief
colorama
yarl
chardet
libneko
psutil
flask
humanize
discord_webhook
DiscordUtils
discord-components
discord-py-slash-command
discord.py==1.7.0